requirements.txt
ruamel.yaml>=0.17.21
sanic>=21.3.4
pydantic>=1.9.0
locust>=2.2.3